refactor(ff-decode): move top-level utility files into shared/ module
