+++ crates/ff-decode/src/shared/hardware.rs
