+++ crates/ff-decode/src/shared/mod.rs
@@ -1,0 +1,8 @@
+//! Shared utility types for ff-decode.
+
+mod hardware;
+pub(crate) mod network;
+mod seek;
+
+pub use hardware::HardwareAccel;
+pub use seek::SeekMode;
