+++ crates/ff-decode/src/shared/network.rs
