+++ crates/ff-decode/src/shared/seek.rs
